requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
statsmodels
python-binance
websockets
copulas
pycopula
scipy
ccxt
 
